[#350] remove old dom visu
