--- frontend/app/src/components/hexagonalBoard/HexagonalBoard.css
@@ -1,26 +1,0 @@
-.board {
-	will-change: transform;
-	display: flex;
-	flex-direction: column;
-	flex-wrap: nowrap;
-	transform: scale(1);
-	transform-origin: top left;
-}
-
-.boardRow {
-	display: flex;
-	flex-direction: row;
-	width: fit-content;
-	height: fit-content;
-	gap: 10px;
-}
-
-.boardRow.even {
-	margin-left: 0;
-	margin-top: -70px;
-}
-
-.boardRow.odd {
-	margin-left: 140px;
-	margin-top: -70px;
-}
