--- frontend/app/src/components/hexagonalTile/HexagonalTile.css
@@ -1,13 +1,0 @@
-.textContent {
-	font-size: 192px;
-	font-weight: bolder;
-	position: absolute;
-	left: 50%;
-	top: 50%;
-	transform: translate(-50%, -50%);
-	color: white;
-}
-
-.layoutedHex {
-	float: left;
-}
