+++ frontend/app/src/domain/HexagonalTileProps.tsx
@@ -1,0 +1,9 @@
+export interface HexagonalTileProps {
+	fill: string;
+	isDimmed: boolean;
+	stroke: string;
+	strokeWidth: string;
+	identifier: string;
+	textContent: string;
+	textColor: string;
+}
